bk/pyproject.toml
[project]
name = "shiori-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "ebooklib>=0.20",
    "mysqlclient>=2.2.8",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
